Split the project into modules
