+++ src/chars.rs
@@ -1,0 +1,81 @@
+pub const HLINE: char = '\u{c4}';
+pub const VLINE: char = '\u{b3}';
+pub const NE: char = '\u{bf}';
+pub const NW: char = '\u{da}';
+pub const SE: char = '\u{d9}';
+pub const SW: char = '\u{c0}';
+pub const TEEW: char = '\u{b4}';
+pub const TEEE: char = '\u{c3}';
+pub const TEEN: char = '\u{c1}';
+pub const TEES: char = '\u{c2}';
+pub const CROSS: char = '\u{c5}';
+pub const DHLINE: char = '\u{cd}';
+pub const DVLINE: char = '\u{ba}';
+pub const DNE: char = '\u{bb}';
+pub const DNW: char = '\u{c9}';
+pub const DSE: char = '\u{bc}';
+pub const DSW: char = '\u{c8}';
+pub const DTEEW: char = '\u{b9}';
+pub const DTEEE: char = '\u{cc}';
+pub const DTEEN: char = '\u{ca}';
+pub const DTEES: char = '\u{cb}';
+pub const DCROSS: char = '\u{ce}';
+pub const BLOCK1: char = '\u{b0}';
+pub const BLOCK2: char = '\u{b1}';
+pub const BLOCK3: char = '\u{b2}';
+pub const ARROW_N: char = '\u{18}';
+pub const ARROW_S: char = '\u{19}';
+pub const ARROW_E: char = '\u{1a}';
+pub const ARROW_W: char = '\u{1b}';
+pub const ARROW2_N: char = '\u{1e}';
+pub const ARROW2_S: char = '\u{1f}';
+pub const ARROW2_E: char = '\u{10}';
+pub const ARROW2_W: char = '\u{11}';
+pub const DARROW_H: char = '\u{1d}';
+pub const DARROW_V: char = '\u{12}';
+pub const CHECKBOX_UNSET: char = '\u{e0}';
+pub const CHECKBOX_SET: char = '\u{e1}';
+pub const RADIO_UNSET: char = '\u{09}';
+pub const RADIO_SET: char = '\u{0a}';
+pub const SUBP_NW: char = '\u{e2}';
+pub const SUBP_NE: char = '\u{e3}';
+pub const SUBP_N: char = '\u{e4}';
+pub const SUBP_SE: char = '\u{e5}';
+pub const SUBP_DIAG: char = '\u{e6}';
+pub const SUBP_E: char = '\u{e7}';
+pub const SUBP_SW: char = '\u{e8}';
+pub const SMILIE: char = '\u{01}';
+pub const SMILIE_INV: char = '\u{02}';
+pub const HEART: char = '\u{03}';
+pub const DIAMOND: char = '\u{04}';
+pub const CLUB: char = '\u{05}';
+pub const SPADE: char = '\u{06}';
+pub const BULLET: char = '\u{07}';
+pub const BULLET_INV: char = '\u{08}';
+pub const MALE: char = '\u{0b}';
+pub const FEMALE: char = '\u{0c}';
+pub const NOTE: char = '\u{0d}';
+pub const NOTE_DOUBLE: char = '\u{0e}';
+pub const LIGHT: char = '\u{0f}';
+pub const EXCLAM_DOUBLE: char = '\u{13}';
+pub const PILCROW: char = '\u{14}';
+pub const SECTION: char = '\u{15}';
+pub const POUND: char = '\u{9c}';
+pub const MULTIPLICATION: char = '\u{9e}';
+pub const FUNCTION: char = '\u{9f}';
+pub const RESERVED: char = '\u{a9}';
+pub const HALF: char = '\u{ab}';
+pub const ONE_QUARTER: char = '\u{ac}';
+pub const COPYRIGHT: char = '\u{b8}';
+pub const CENT: char = '\u{bd}';
+pub const YEN: char = '\u{be}';
+pub const CURRENCY: char = '\u{cf}';
+pub const THREE_QUARTERS: char = '\u{f3}';
+pub const DIVISION: char = '\u{f6}';
+pub const GRADE: char = '\u{f8}';
+pub const UMLAUT: char = '\u{f9}';
+pub const POW1: char = '\u{fb}';
+pub const POW3: char = '\u{fc}';
+pub const POW2: char = '\u{fd}';
+pub const BULLET_SQUARE: char = '\u{fe}';
+
